Fix a couple bugs in the stack management script
There are two use cases for a stack restart,
one where you want to take a node down for maintentance,
and two where you really want to stop the whole monasca
stack.  monasca-thresh is a cluster-wide process and can
be stopped from any node in the cluster.  But stopping it
when taking a single node out of service isn't appropriate.

Also update the lag checks for the new/correct way to
get kafka consumer lag.

Change-Id: I333b791c903a45e8a0fc101ade229991a03da2cc